requirements.txt
fastapi == 0.85.0
uvicorn == 0.18.3
pydantic == 1.10.2
pymongo == 4.2.0
python-dotenv == 0.21.0
dnspython == 2.2.1